+++ hutool-core/src/test/java/cn/hutool/core/net/UrlQueryTest.java
@@ -26,4 +26,12 @@ public class UrlQueryTest {
 		final String queryStr = urlBuilder.getQueryStr();
 		Assert.assertEquals("imageMogr2/auto-orient/thumbnail/500&pid=259848", queryStr);
 	}
+
+	@Test
+	public void parseTest2(){
+		String requestUrl = "http://192.168.1.1:8080/pc?=d52i5837i4ed=o39-ap9e19s5--=72e54*ll0lodl-f338868d2";
+		UrlQuery q = new UrlQuery();
+		UrlQuery parse = q.parse(requestUrl, Charset.defaultCharset());
+		Assert.assertEquals("=d52i5837i4ed=o39-ap9e19s5--=72e54*ll0lodl-f338868d2", parse.toString());
+	}
 }